agent/rag_agent/nodes/generate.py
from langchain import hub
from langchain_core.output_parsers import StrOutputParser
from langchain_groq import ChatGroq
from dotenv import load_dotenv

load_dotenv()

prompt = hub.pull("rlm/rag-prompt")

llm = ChatGroq(model_name="llama-3.3-70b-versatile", temperature=0)


def format_docs(docs):
    return "\n\n".join(doc.page_content for doc in docs)


rag_chain = prompt | llm | StrOutputParser()
